fix(review): dict-payload guard, extract discovery_judge module, multi-token collision disambiguation (#2 #3 #5)
